(PC-15716)[API] refactor: move subscription scripts to scripts/subscription
