+++ api/src/pcapi/scripts/subscription/dms/handle_deleted_dms_applications.py
