+++ api/tests/scripts/beneficiary/archive_dms_applications_test.py
@@ -6,7 +6,7 @@ from pcapi.connectors.dms import api as api_dms
 from pcapi.core import testing
 from pcapi.core.fraud import factories as fraud_factories
 from pcapi.core.fraud import models as fraud_models
-from pcapi.scripts.beneficiary import archive_dms_applications
+from pcapi.scripts.subscription.dms import archive_dms_applications
 
 from tests.scripts.beneficiary.fixture import make_parsed_graphql_application
